# backend/static/book-scents.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/api/v1/login": {
      "post": {
        "summary": "登录",
        "deprecated": false,
        "description": "",
        "tags": [
          "用户"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "title": "用户名"
                  },
                  "password": {
                    "type": "string",
                    "title": "密码"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "title": "JWT 认证"
                    },
                    "profile": {
                      "$ref": "#/components/schemas/%E7%94%A8%E6%88%B7%E6%A8%A1%E5%9E%8B"
                    }
                  },
                  "required": [
                    "token",
                    "profile"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/register": {
      "post": {
        "summary": "注册",
        "deprecated": false,
        "description": "",
        "tags": [
          "用户"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "title": "用户名"
                  },
                  "password": {
                    "type": "string",
                    "title": "用户密码"
                  },
                  "mobile": {
                    "type": "string",
                    "title": "手机号"
                  }
                },
                "required": [
                  "username",
                  "password",
                  "mobile"
                ]
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profile": {
                      "$ref": "#/components/schemas/%E7%94%A8%E6%88%B7%E6%A8%A1%E5%9E%8B"
                    },
                    "token": {
                      "type": "string",
                      "title": "jwt 认证"
                    }
                  },
                  "required": [
                    "profile",
                    "token"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/books": {
      "get": {
        "summary": "书籍列表",
        "deprecated": false,
        "description": "",
        "tags": [
          "书籍"
        ],
        "parameters": [
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageNum",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "DESC",
                "ASC"
              ]
            }
          },
          {
            "name": "filed",
            "in": "query",
            "description": "配合 order 进行排序",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "HOT",
                "RECOMMEND"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "list": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/%E4%B9%A6%E7%B1%8D%E4%BF%A1%E6%81%AF"
                      }
                    },
                    "total": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "list",
                    "total"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "上传书籍",
        "deprecated": false,
        "description": "",
        "tags": [
          "书籍"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "description": "书籍名",
                    "example": "",
                    "type": "string"
                  },
                  "author": {
                    "description": "作者",
                    "example": "",
                    "type": "string"
                  },
                  "cover": {
                    "description": "书籍封面",
                    "example": "",
                    "type": "string"
                  },
                  "description": {
                    "description": "书籍描述",
                    "example": "",
                    "type": "string"
                  },
                  "url": {
                    "description": "阅读地址",
                    "example": "",
                    "type": "string"
                  }
                }
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/books/{id}": {
      "get": {
        "summary": "书籍详情",
        "deprecated": false,
        "description": "",
        "tags": [
          "书籍"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "title": "书籍id"
                    },
                    "author": {
                      "type": "string",
                      "title": "作者"
                    },
                    "name": {
                      "type": "string",
                      "title": "书籍名"
                    },
                    "cover": {
                      "type": "string",
                      "title": "书籍封面"
                    },
                    "description": {
                      "type": "string",
                      "title": "书籍描述"
                    },
                    "rating": {
                      "type": "integer",
                      "description": "1-5 之前数字允许出现.5",
                      "title": "评分：1-5 之前数字允许出现.5"
                    },
                    "statement": {
                      "type": "array",
                      "items": {
                        "type": "integer"
                      },
                      "description": "1 星到 5 星之前的百分比数据",
                      "title": "根据评分计算出的1-5星之间的百分比数据"
                    },
                    "url": {
                      "type": "string",
                      "title": "书籍地址"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "cover",
                    "description",
                    "rating",
                    "statement",
                    "author",
                    "url"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/books/{id}/read": {
      "patch": {
        "summary": "记录阅读次数",
        "deprecated": false,
        "description": "",
        "tags": [
          "书籍"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}